requirements.txt
fastapi
uvicorn
aiohttp
aiomysql
mysql-connector-python
openai
jinja2
python-dotenv
beautifulsoup4
requests
